+++ LinkEngine/WorldGen/Biome.cs
@@ -7,8 +7,9 @@ namespace LinkEngine.WorldGen
         public int ID { get; set; }
         public string Name { get; set; }
         public List<Tile> AvailibleTiles { get; set; }
+        public List<Biome> RelatedBiomes { get; set; }
 
-        public Biome(int id, string name, Tile[] tileAry)
+        public Biome(int id, string name, Tile[] tileAry, Biome[] biomeAry)
         {
             ID = id;
             Name = name;
@@ -17,6 +18,10 @@ namespace LinkEngine.WorldGen
             for(int i = 0; i < tileAry.Length; i++)
             {
                 AvailibleTiles.Add(tileAry[i]);
+            }
+            for (int i = 0; i < biomeAry.Length; i++)
+            {
+                RelatedBiomes.Add(biomeAry[i]);
             }
         }
     }